Adding snap tools
The two scripts snap.sh and snap-set.sh are
useful to get the snapshot of the current
git state.

snap.sh is getting the snapshot of the
gits that are in the current directory
and write it to its standard output.

snap-set.sh is reading a snapshot on
its input and (re)create the output
directories as it was for the snapshot.

Change-Id: Ieae37eb13aee890f0b020001c5771ebf6d3c918a